(#189) Introduce something_print
