+++ src/something_commands.cpp
@@ -28,10 +28,6 @@ void command_close(Game *game, String_View)
     game->console.toggle();
 }
 
-void sprint1(String_Buffer *sbuffer, SDL_Color color)
-{
-    sprint(sbuffer, "{", color.r, ",", color.g, ",", color.b, ",", color.a, "}");
-}
 
 #ifndef SOMETHING_RELEASE
 void command_set(Game *game, String_View args)
@@ -93,15 +89,3 @@ void command_reload(Game *game, String_View)
 }
 
 #endif // SOMETHING_RELEASE
-
-void sprint1(String_Buffer *sbuffer, Particles::State state)
-{
-    switch (state) {
-    case Particles::DISABLED:
-        sprint(sbuffer, "DISABLED");
-        break;
-    case Particles::EMITTING:
-        sprint(sbuffer, "EMITTING");
-        break;
-    }
-}
