+++ src/something_main.cpp
@@ -3,12 +3,6 @@
 const int SIMULATION_FPS = 60;
 const float SIMULATION_DELTA_TIME = 1.0f / SIMULATION_FPS;
 
-template <typename T>
-void print1(FILE *stream, Vec2<T> v)
-{
-    print(stream, '(', v.x, ',', v.y, ')');
-}
-
 Game game = {};
 
 int main(void)
